refactor(workflows): split the inverse engine along the seams it already had
workflows/inverse.py was 1,445 lines and roughly fifty top-level definitions, but its parts were
not tangled — the cross-references between them measure as a strict hierarchy with no cycles at
all: config <- seeds <- {paths, report} <- engine. The package is that measurement, one module per
layer, so the boundaries are the ones the code already obeyed rather than ones invented for it.

    config.py   513   the TOML schema: targets, scenarios, paths, seeds, and their builders
    report.py   417   figures and markdown — the leaf everything else feeds
    engine.py   225   run: the scenario loop, composing the other four
    seeds.py    202   which compositions a scenario starts from
    paths.py    202   running one seed set down one path

Verified as a relocation, not a rewrite: all 48 top-level definitions are token-identical to the
originals, none lost and none added, and all 14 module constants keep their values.

That check is in the commit message because the first attempt failed it. Slicing a class by its
`lineno` starts at `class`, not at the decorator above it, so five config dataclasses arrived as
plain classes — the same near-miss this PR already hit once, when `evaluate_targets` nearly lost
its @torch.no_grad(). The tests caught it immediately; the diff would not have.

The re-export list is the public API only. Eight private helpers that inverse_test.py imported
through the module now come from the module that owns them: widening a package's surface to spare
a test one import line would say those helpers are API when they are not.
